# api/swagger/v1/team.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "team service",
    "version": "1.0",
    "contact": {
      "name": "team service",
      "url": "https://github.com/ckbball/dev-team",
      "email": "blah@happy.org"
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/teams": {
      "post": {
        "operationId": "CreateTeam",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/teamTeamUpsertResponse"
            }
          },
          "404": {
            "description": "Returned when resource doesn't exist",
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/teamTeamUpsertRequest"
            }
          }
        ],
        "tags": [
          "TeamService"
        ]
      }
    },
    "/v1/teams/project/{id}": {
      "post": {
        "operationId": "UpsertTeamProject",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/teamProjectUpsertResponse"
            }
          },
          "404": {
            "description": "Returned when resource doesn't exist",
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/teamProjectUpsertRequest"
            }
          }
        ],
        "tags": [
          "TeamService"
        ]
      }
    },
    "/v1/teams/{id}": {
      "delete": {
        "operationId": "DeleteTeam",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/teamTeamDeleteResponse"
            }
          },
          "404": {
            "description": "Returned when resource doesn't exist",
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "api",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "TeamService"
        ]
      }
    },
    "/v1/teams/{id}/members": {
      "post": {
        "operationId": "AddMember",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/teamMemberUpsertResponse"
            }
          },
          "404": {
            "description": "Returned when resource doesn't exist",
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/teamMemberUpsertRequest"
            }
          }
        ],
        "tags": [
          "TeamService"
        ]
      }
    },
    "/v1/teams/{id}/members/{member_number}": {
      "delete": {
        "operationId": "RemoveMember",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/teamMemberDeleteResponse"
            }
          },
          "404": {
            "description": "Returned when resource doesn't exist",
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "member_number",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "api",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "TeamService"
        ]
      }
    }
  },
  "definitions": {
    "teamMemberDeleteResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "count": {
          "type": "string"
        }
      }
    },
    "teamMemberUpsertRequest": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "member_id": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      }
    },
    "teamMemberUpsertResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "member_number": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
    "teamProject": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "languages": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "name": {
          "type": "string"
        },
        "github_link": {
          "type": "string"
        }
      }
    },
    "teamProjectUpsertRequest": {
      "type": "object",
      "properties": {
        "project": {
          "$ref": "#/definitions/teamProject"
        },
        "api": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    },
    "teamProjectUpsertResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
    "teamTeam": {
      "type": "object",
      "properties": {
        "leader": {
          "type": "string"
        },
        "members": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "name": {
          "type": "string"
        },
        "open_roles": {
          "type": "integer",
          "format": "int32"
        },
        "skills": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "size": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "teamTeamDeleteResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "count": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    },
    "teamTeamUpsertRequest": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "team": {
          "$ref": "#/definitions/teamTeam"
        }
      }
    },
    "teamTeamUpsertResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    }
  }
}